feat: implement Step 3 - core engines for universal object system
Implemented all core engines to make TDD tests pass:

- GameWorld: Immutable game state management with object storage,
  property/state updates, parent relationships, ID generation, and queries

- TargetResolver: Object targeting system supporting resolution by ID, type,
  relation (self/opponent), property matching, and context-aware "this" refs

- TriggerEngine: Trigger condition matching with object type filtering,
  property change detection, specific value matching, object matchers,
  and custom condition expression evaluation

- EffectEngine: Effect execution supporting property modification (delta/absolute),
  object movement, creation with auto-generated IDs, destruction, and
  sequential effect processing

- Updated test files to use real implementations instead of TODO placeholders
- Added SimpleEngineTest for basic verification of core functionality
- Fixed JavaScript export compatibility (ObjectIdResult vs Pair, Int vs Long)
- All code compiles successfully for both JVM and JavaScript targets

+++ catenin/src/commonMain/kotlin/org/junction/catenin/model/GameObject.kt
@@ -45,7 +45,7 @@ data class PropertyChangeEvent(
     val propertyPath: String, // e.g., "properties.health" or "states.tapped"
     val oldValue: PropertyValue?,
     val newValue: PropertyValue,
-    val timestamp: Long = 0L // for ordering/replay
+    val timestamp: Int = 0 // for ordering/replay - using Int instead of Long for JS compatibility
 )
 
 /**
